Add macroexpand to emacs-lisp-hydra
-----------

emacs-lisp-macroexpand is an interactive compiled Lisp function in
‘elisp-mode.el’.

(emacs-lisp-macroexpand)

Macroexpand the form after point.
Comments in the form will be lost.

-----------

macroexpand is a built-in function in ‘C source code’.

(macroexpand FORM &optional ENVIRONMENT)

Return result of expanding macros at top level of FORM.
If FORM is not a macro call, it is returned unchanged.
Otherwise, the macro is expanded and the expansion is considered
in place of FORM.  When a non-macro-call results, it is returned.

The second optional arg ENVIRONMENT specifies an environment of macro
definitions to shadow the loaded ones for use in file byte-compilation.