[frame-workflow] Fix up again
The previous implementation somehow (why?) didn't show the two
(emacs.org and magit) buffers as expected, but this seems to work
well.

Note that, if you first close the sidebar window and then quit magit,
the previous window layout is restored. This is awkward, but I cannot
find a solution for now.