[org-make-toc] Add support for org-make-toc buffer-local variable
One annoying issue with org-make-toc was that I have to add it to
before-save-hook and that it is an unsafe variable.
Because it cannot be added to the list of safe variables in my
custom file, it prompts for confirmation every time I load a file
that includes a setting for org-make-toc.

I resolved issue by adding a buffer-local variable named org-make-toc
as well as a hook in org-mode which turns on org-make-toc-mode.

Now my emacs.org contains `org-make-toc: t` local variable instead of
the before-save-hook, and it can be added to safe variables.