[dracula-theme] Set custom face for hl-line
Because the default hl-line face makes the background of
an active line lighter, which makes the text less readable.
I temporarily set the background color to black, which is
darker than the default background of the theme.

I know this is weird, but I can't think of a better face
setting for hl-line.