[ivy] Disable ivy-posframe and set the height to 10
There are some corner cases where ivy-posframe doesn't render well.
The package is not mature enough at present, so I'll stop using it for
now. I would like to use it when it's properly implemented, though.

Another reason I've switched back to the normal display function is
that the minibuffer window is displayed very fast on Pixelbook.
It was slow on my previous laptop, which made me use ivy-posframe.