Remap auto yasnippet commands on register map and make C-c a & e free
I defined a key translation which translates M-r to C-x r, which is a
prefix for register and rectangle commands. auto-yasnippet is somewhat
analogous to registers, so I remapped the commands in the package.
Use M-r for recalling something in Emacs.

Now C-c a and C-c e are free, so I can use them for other purposes. I
will use C-c e as a prefix for eval commands, as quite a few language
modes supports such a feature.