Add Japanese mozc
It turns out that mozc.el can be used without running fcitx or ibus.

It only depends on mozc_emacs_helper and the mozc server, which can
be installed as independent packages on Debian.
I'm not sure about other distros. Apparently, it is unavailable in
nixpkgs yet, so I would like to contribute to it in the future.

Note that it takes huge time to clone the repository of mozc,
so I've added the recent version of mozc.el to contrib directory
to avoid installing it using straight.el.

I created katawa.el https://github.com/akirak/katawa.el in the past,
but mozc-temp seems to be way better if you have mozc_emacs_helper
installed.