[defaults] Allow disabling global-auto-revert-mode
On WSL where filesystem IO is much slower than native Linux,
autorevert is fatal to performance. I want global-auto-revert-mode to
be the default on most machines, but I also want to allow disabling
the feature, so I made this change.

With the change, you can disable the global mode by overriding the
variable in custom-file.